backend/requirements.txt
# 后端框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 任务队列
celery==5.3.4
redis==5.0.1
flower==2.0.1

# LangChain相关
langchain==0.0.335
langchain-community==0.0.6
langchain-openai==0.0.2
chromadb==0.4.17
openai==1.3.5

# 通义千问
dashscope==1.14.1

# 文档处理
PyMuPDF==1.23.8
pypdf==3.17.1
python-docx==1.1.0

# 中文分词
jieba==0.42.1

# 工具库
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.3
tiktoken==0.5.1
requests==2.31.0

# 监控和日志
prometheus-client==0.19.0
structlog==23.2.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0 